# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Proyecto Android Laravel",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api"
        }
    ],
    "paths": {
...
        "/api/listaUsuarios": {
            "get": {
                "summary": "listaUsuarios",
                "responses": {
                    "200": {
                        "description": "Se actualizan los datos del usuario logeado."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/listaPlatos/{tiempoToken}": {
            "get": {
                "description": "LIstar los platos",
                "operationId": "listaPlatos",
                "parameters": [
                    {
                        "name": "tiempoToken",
                        "in": "path",
                        "description": "Tiempo del token",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Ha ocurrido un error."
                    },
                    "200": {
                        "description": "Detalles platos."
                    }
                }
            }
        },
        "/crearPlato": {
            "post": {
                "description": "Crear platos",
                "operationId": "crearPlato",
                "parameters": [
                    {
                        "name": "nombre",
                        "in": "path",
                        "description": "Nombre del plato",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "String"
                        }
                    },
                    {
                        "name": "precio",
                        "in": "path",
                        "description": "Precio del plato",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "idCreador",
                        "in": "path",
                        "description": "Persona que creo el plato",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Ha ocurrido un error."
                    },
                    "200": {
                        "description": "Se ha creado el plato."
                    }
                }
            }
        },
        "/actualizarPlato": {
            "put": {
                "description": "Actualizar platos",
                "operationId": "actualizarPlato",
                "parameters": [
                    {
                        "name": "nombre",
                        "in": "path",
                        "description": "Nombre del plato",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "String"
                        }
                    },
                    {
                        "name": "precio",
                        "in": "path",
                        "description": "Precio del plato",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del plato",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Ha ocurrido un error."
                    },
                    "200": {
                        "description": "El plato se ha actualizado."
                    }
                }
            }
        },
        "/dashabilitatPlato": {
            "put": {
                "description": "Deshabilitar plato",
                "operationId": "dashabilitatPlato",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del plato",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Ha ocurrido un error."
                    },
                    "200": {
                        "description": "PLato deshabilitado."
                    }
                }
            }
        },
...
    }
}